+++ changelog.md
@@ -1,6 +1,6 @@
 ### Revision History
 * 1.68.0
-  * Fixed `UniqueIdGenerator` code that determines a unique `serverId`.  
+  * Fixed: `UniqueIdGenerator` now correctly gets last two digits of ID using 3 attempts - JAVA_UTIL_CLUSTERID (optional), CF_INSTANCE_INDEX, and finally using SecuritRandom for the last two digits.  
   * Removed `log4j` in favor of `slf4j` and `logback`.
 * 1.67.0
   * Updated log4j dependencies to version `2.17.1`.